feat: implement more activity features
The importer now folds some data coming in from the FIT session file
into the activity. This way we can keep the database structure
clean and simple.

+++ db/migrations/20200524083549_create_activity_records.up.sql
@@ -2,8 +2,8 @@ CREATE TABLE activity_records
 (
   id BIGSERIAL PRIMARY KEY,
   "timestamp" TIMESTAMP NOT NULL,
-  position_lat INTEGER,
-  position_long INTEGER,
+  position_lat FLOAT,
+  position_long FLOAT,
   distance FLOAT,
   time_from_course INTEGER,
   compressed_speed_distance bytea,
